src/third_party/blink/web_tests/html/document_metadata/referrer-outside-head-use-counter.html

<!DOCTYPE html>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<body>
  <meta name=referrer content=origin />
  <script>
    'use strict';

    test(() => {
      const kHTMLMetaElementReferrerPolicyOutsideHead = 3314; // from web_feature.mojom
      const kHTMLMetaElementReferrerPolicyMultipleTokens = 3315;
      assert_true(internals.isUseCounted(document, kHTMLMetaElementReferrerPolicyOutsideHead));
      assert_false(internals.isUseCounted(document, kHTMLMetaElementReferrerPolicyMultipleTokens));
    }, '<meta name=referrer> outside <head> is use counted');
  </script>
